requirements.txt
crawl4ai==0.4.247
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.10.6
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
lxml~=5.3
fake-useragent==1.4.0
tqdm==4.66.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
